+++ strelyaev.roman/T1/inputShapes.hpp
@@ -1,9 +1,6 @@
 #ifndef INPUT_SHAPES_HPP
 #define INPUT_SHAPES_HPP
 #include "shape.hpp"
-#include "rectangle.hpp"
-#include "triangle.hpp"
-#include "Parallelogram.hpp"
 namespace strelyaev
 {
   Shape * inputRectangle(const char string[]);
