+++ strelyaev.roman/T1/parallelogram.hpp
